refactor(frontend): extract SettingsShellLayout for reuse
The /settings/* nav shell (header + left-nav + footer) is about to
serve two routes: the existing settings tabs and the new
/me/profile model-of-me page (#752). Pull the shell into a shared
client component so each route's layout.tsx is a thin wrapper and
the nav items list stays the single source of truth.

Adds the "Your model" nav entry (en + es) that points at
/me/profile so the model-of-me page is discoverable from any
settings tab.
