+++ apps/frontend/locales/en/settings.json
@@ -2,6 +2,7 @@
   "nav": {
     "settings": "Settings",
     "profile": "Profile",
+    "modelOfMe": "Your model",
     "addresses": "Addresses",
     "notifications": "Notifications",
     "privacy": "Privacy",
