+++ apps/frontend/locales/es/settings.json
@@ -2,6 +2,7 @@
   "nav": {
     "settings": "Configuracion",
     "profile": "Perfil",
+    "modelOfMe": "Tu modelo",
     "addresses": "Direcciones",
     "notifications": "Notificaciones",
     "privacy": "Privacidad",
